refactor(core): move useDebounce to a utility hook
