+++ packages/core/src/components/FormBuilder/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './useFormCustomizationState';
 export * from './useContractWidgetState';
 export * from './useFormBuilderState';
 export * from './useCompleteStepState';
+export * from './useDebounce';
